reorganize packages for test fixtures
