+++ avs-api/src/test/java/org/example/age/avs/api/AvsApiTest.java
@@ -9,10 +9,10 @@ import java.io.IOException;
 import java.util.Map;
 import javax.inject.Singleton;
 import okhttp3.Response;
-import org.example.age.avs.api.test.StubAvsServiceModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
-import org.example.age.test.server.undertow.TestUndertowModule;
+import org.example.age.test.avs.service.StubAvsServiceModule;
+import org.example.age.test.common.server.undertow.TestUndertowModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestUndertowServer;
 import org.junit.jupiter.api.Test;
