+++ site-api/src/test/java/org/example/age/site/api/SiteApiTest.java
@@ -18,8 +18,8 @@ import org.example.age.data.crypto.BytesValue;
 import org.example.age.data.crypto.DigitalSignature;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.VerifiedUser;
-import org.example.age.site.api.test.StubSiteServiceModule;
-import org.example.age.test.server.undertow.TestUndertowModule;
+import org.example.age.test.common.server.undertow.TestUndertowModule;
+import org.example.age.test.site.service.StubSiteServiceModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestUndertowServer;
 import org.junit.jupiter.api.Test;
