+++ site-api/src/testFixtures/java/org/example/age/test/site/service/StubSiteServiceModule.java
@@ -1,4 +1,4 @@
-package org.example.age.site.api.test;
+package org.example.age.test.site.service;
 
 import dagger.Binds;
 import dagger.Module;
@@ -6,7 +6,7 @@ import org.example.age.common.service.data.DataMapperModule;
 import org.example.age.common.service.data.UserAgentAuthMatchDataExtractorModule;
 import org.example.age.site.api.SiteApi;
 import org.example.age.site.api.SiteApiModule;
-import org.example.age.test.service.data.TestAccountIdExtractorModule;
+import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 
 /** Dagger module that binds dependencies for <code>@Named("api") HttpHandler</code>. */
 @Module(
