+++ site-service/src/test/java/org/example/age/site/service/SiteServiceTest.java
@@ -25,9 +25,9 @@ import org.example.age.data.crypto.SigningKeys;
 import org.example.age.site.api.AvsLocation;
 import org.example.age.site.service.config.SiteConfig;
 import org.example.age.site.service.store.InMemoryVerificationStoreModule;
-import org.example.age.site.service.test.FakeAvsServiceModule;
-import org.example.age.test.server.undertow.TestUndertowModule;
-import org.example.age.test.service.data.TestAccountIdExtractorModule;
+import org.example.age.test.avs.service.FakeAvsServiceModule;
+import org.example.age.test.common.server.undertow.TestUndertowModule;
+import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 import org.example.age.testing.client.TestClient;
 import org.example.age.testing.server.TestUndertowServer;
 import org.junit.jupiter.api.BeforeAll;
