+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsService.java
@@ -1,4 +1,4 @@
-package org.example.age.site.service.test;
+package org.example.age.test.avs.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
