+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsServiceModule.java
@@ -1,4 +1,4 @@
-package org.example.age.site.service.test;
+package org.example.age.test.avs.service;
 
 import dagger.Binds;
 import dagger.Module;
@@ -9,7 +9,7 @@ import org.example.age.avs.api.SiteLocation;
 import org.example.age.common.service.data.DataMapperModule;
 import org.example.age.common.service.data.DisabledAuthMatchDataExtractorModule;
 import org.example.age.infra.service.client.RequestDispatcherModule;
-import org.example.age.test.service.data.TestAccountIdExtractorModule;
+import org.example.age.test.common.service.data.TestAccountIdExtractorModule;
 
 /**
  * Dagger module that binds dependencies for <code>@Named("api") HttpHandler</code>.
